requirements.txt
ntscraper
requests
python-dotenv
